+++ src/app/item-popup/ItemSocketsWeapons.m.scss.d.ts
@@ -1,0 +1,10 @@
+// This file is automatically generated.
+// Please do not change this file!
+interface CssExports {
+  'archetype': string;
+  'archetypeMod': string;
+  'perks': string;
+  'row': string;
+}
+export const cssExports: CssExports;
+export default cssExports;
